clean up / refactor for style
One improvement this suite should have over my old stuff is that the
--next functionality of nssh can be exposed by 'gdsh run --next' to
do single-stepping through a list running a script or command. I've
wanted this a few times and usually drop back to shell loops, which
are not as safe or expressive, not to mention a lot more typing.